+++ src/parallel/DistributedObject.h
@@ -33,9 +33,9 @@ struct DistributedObject
 
     void wait();
 
-    void add_send_link(size_t id, const nTuple<ptrdiff_t, 3> &offset, const data_model::DataSet *);
+    void add_send_link(size_t id, const nTuple<int, 3> &offset, const data_model::DataSet *);
 
-    void add_recv_link(size_t id, const nTuple<ptrdiff_t, 3> &offset, data_model::DataSet *);
+    void add_recv_link(size_t id, const nTuple<int, 3> &offset, data_model::DataSet *);
 
     bool is_ready() const;
